+++ src/net/sf/mpxj/primavera/PrimaveraPMFileReader.java
@@ -25,7 +25,6 @@ package net.sf.mpxj.primavera;
 
 import java.io.InputStream;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -47,7 +46,6 @@ import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
 
 import net.sf.mpxj.AssignmentField;
-import net.sf.mpxj.ChildTaskContainer;
 import net.sf.mpxj.ConstraintType;
 import net.sf.mpxj.CustomFieldContainer;
 import net.sf.mpxj.DateRange;
@@ -421,17 +419,20 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
    {
       List<WBSType> wbs = project.getWBS();
       List<ActivityType> tasks = project.getActivity();
-
       Set<Integer> uniqueIDs = new HashSet<Integer>();
+      Set<Task> wbsTasks = new HashSet<Task>();
 
       //
       // Read WBS entries and create tasks
       //
+      Collections.sort(wbs, WBS_ROW_COMPARATOR);
+
       for (WBSType row : wbs)
       {
          Task task = m_projectFile.addTask();
          Integer uniqueID = row.getObjectId();
          uniqueIDs.add(uniqueID);
+         wbsTasks.add(task);
 
          task.setUniqueID(uniqueID);
          task.setName(row.getName());
@@ -592,70 +593,10 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
          m_eventManager.fireTaskReadEvent(task);
       }
 
-      sortActivities(TaskField.TEXT1, m_projectFile);
+      new ActivitySorter(TaskField.TEXT1, wbsTasks).sort(m_projectFile);
+
       updateStructure();
       updateDates();
-   }
-
-   /**
-    * Ensure activities are sorted into Activity ID order to match Primavera.
-    *
-    * @param activityIDField field containing the Activity ID value
-    * @param container object containing the tasks to process
-    */
-   private void sortActivities(final FieldType activityIDField, ChildTaskContainer container)
-   {
-      // Do we have any tasks?
-      List<Task> tasks = container.getChildTasks();
-      if (!tasks.isEmpty())
-      {
-         for (Task task : tasks)
-         {
-            //
-            // Sort child activities
-            //
-            sortActivities(activityIDField, task);
-
-            //
-            // Sort Order:
-            // 1. Activities come first
-            // 2. WBS come last
-            // 3. Activities ordered by activity ID
-            // 4. WBS ordered by ID
-            //
-            Collections.sort(tasks, new Comparator<Task>()
-            {
-               @Override public int compare(Task t1, Task t2)
-               {
-                  boolean t1HasChildren = !t1.getChildTasks().isEmpty();
-                  boolean t2HasChildren = !t2.getChildTasks().isEmpty();
-
-                  // Both are WBS
-                  if (t1HasChildren && t2HasChildren)
-                  {
-                     return t1.getID().compareTo(t2.getID());
-                  }
-
-                  // Both are activities
-                  if (!t1HasChildren && !t2HasChildren)
-                  {
-                     String activityID1 = (String) t1.getCurrentValue(activityIDField);
-                     String activityID2 = (String) t2.getCurrentValue(activityIDField);
-
-                     if (activityID1 == null || activityID2 == null)
-                     {
-                        return (activityID1 == null && activityID2 == null ? 0 : (activityID1 == null ? 1 : -1));
-                     }
-
-                     return activityID1.compareTo(activityID2);
-                  }
-
-                  // One activity one WBS
-                  return t1HasChildren ? 1 : -1;
-               }
-            });
-         }
-      }
    }
 
    /**
@@ -1091,4 +1032,6 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
       MILESTONE_MAP.put("Finish Milestone", Boolean.TRUE);
       MILESTONE_MAP.put("WBS Summary", Boolean.FALSE);
    }
+
+   private static final WbsRowComparatorPMXML WBS_ROW_COMPARATOR = new WbsRowComparatorPMXML();
 }